Minor changes and bug fixes
AddToSystem: corrected errors in conditions of several for loops that
caused charged beads instead of neutral beads to be exchanged.

Config: it now generates coordinates with [0,0,0] in the middle of the box
as expected by dl_meso.

DistrAgg: removed some unnecessary output data (though they're still
calculated if I ever find they're actually useful). The change needs to
be reflected in the manual.