Added sorting functions
SortArray() sorts a simple array of integers (in either ascending or
descendig order).

SortAggStruct() sorts Aggregate structure so that aggregates with a
lower index of the first molecule come first.

SortBonds() sorts 2D array of bonds so that in each bond the first bead
index is lower; all bonds are then sorted so that bonds with a lower
index of the first bead come first.