Changes to AddToSystem
The utility now expands the system by the new beads by default, rather
then switching beads in the original system for new ones; the original
behaviour is preserved via the --switch option. Box size of the new
system can be changed from the original one.

Using --switch option, the old behaviour of switching old beads for new
ones is preserved, while if --switch is not used, the new beads are
added to the system, expanding the new system.